requirements.txt
scapy
mac-vendor-lookup
psutil
pystray
Pillow
cachetools
matplotlib